GeneralSettings popup window class
