+++ .gitignore
@@ -1,3 +1,4 @@
 *.pyc
 .DS_Store
 .idea/
+f-engrave.icon
